+++ garazd_product_label/wizard/product_label.py
@@ -5,10 +5,16 @@ from odoo import api, fields, models, _
 
 class ProductLabel(models.TransientModel):
     _name = "product.label"
+    _description = 'Product Label'
 
-    selected = fields.Boolean('Print', compute='_compute_selected')
-    product_id = fields.Many2one('product.product', 'Product', required=True)
+    selected = fields.Boolean(
+        string='Print',
+        compute='_compute_selected',
+        readonly=True,
+    )
     wizard_id = fields.Many2one('print.product.label', 'Print Wizard')
+    product_id = fields.Many2one('product.product', 'Product', required=True)
+    barcode = fields.Char('Barcode', related='product_id.barcode')
     qty_initial = fields.Integer('Initial Qty', default=1)
     qty = fields.Integer('Label Qty', default=1)
 
@@ -23,10 +29,10 @@ class ProductLabel(models.TransientModel):
     @api.multi
     def action_plus_qty(self):
         for record in self:
-            record.update({'qty': record.qty+1})
+            record.update({'qty': record.qty + 1})
 
     @api.multi
     def action_minus_qty(self):
         for record in self:
             if record.qty > 0:
-                record.update({'qty': record.qty-1})
+                record.update({'qty': record.qty - 1})