+++ garazd_product_label/doc/changelog.rst
@@ -3,7 +3,15 @@
 Changelog
 =========
 
-`1.0.1`
+`12.0.1.0.2`
+-------
+
+- **IMP:** Add to wizard field "output" - to select a type of printing.
+- **IMP:** Add to wizard field "message" - to show messages in the print wizard.
+- **IMP:** Add to label field "barcode" - to show barcode in the wizard.
+- **IMP:** Change the wizard form view: add group names, new fields.
+
+`12.0.1.0.1`
 -------
 
 - Init version